+++ python/cudf/cudf/tests/series/methods/test_pipe.py
@@ -1,0 +1,64 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.
+
+import pandas as pd
+
+import cudf
+from cudf.testing import assert_eq
+from cudf.testing._utils import (
+    assert_exceptions_equal,
+)
+
+
+def test_series_pipe():
+    psr = pd.Series([10, 20, 30, 40])
+    gsr = cudf.Series([10, 20, 30, 40])
+
+    def custom_add_func(sr, val):
+        new_sr = sr + val
+        return new_sr
+
+    def custom_to_str_func(sr, val):
+        new_sr = sr.astype("str") + val
+        return new_sr
+
+    expected = (
+        psr.pipe(custom_add_func, 11)
+        .pipe(custom_add_func, val=12)
+        .pipe(custom_to_str_func, "rapids")
+    )
+    actual = (
+        gsr.pipe(custom_add_func, 11)
+        .pipe(custom_add_func, val=12)
+        .pipe(custom_to_str_func, "rapids")
+    )
+
+    assert_eq(expected, actual)
+
+    expected = (
+        psr.pipe((custom_add_func, "sr"), val=11)
+        .pipe(custom_add_func, val=1)
+        .pipe(custom_to_str_func, "rapids-ai")
+    )
+    actual = (
+        gsr.pipe((custom_add_func, "sr"), val=11)
+        .pipe(custom_add_func, val=1)
+        .pipe(custom_to_str_func, "rapids-ai")
+    )
+
+    assert_eq(expected, actual)
+
+
+def test_series_pipe_error():
+    psr = pd.Series([10, 20, 30, 40])
+    gsr = cudf.Series([10, 20, 30, 40])
+
+    def custom_add_func(sr, val):
+        new_sr = sr + val
+        return new_sr
+
+    assert_exceptions_equal(
+        lfunc=psr.pipe,
+        rfunc=gsr.pipe,
+        lfunc_args_and_kwargs=([(custom_add_func, "val")], {"val": 11}),
+        rfunc_args_and_kwargs=([(custom_add_func, "val")], {"val": 11}),
+    )
